+++ packages/plugin-packages/mashroom-portal/src/backend/dev-server.ts
@@ -32,8 +32,9 @@ setPortalPluginConfig({
     autoExtendAuthentication: false,
     defaultProxyConfig: {},
     ssrConfig: {
-        ssrEnabled: true,
+        ssrEnable: true,
         renderTimoutMs: 2000,
+        cacheEnable: false,
         cacheTTLSec: 300,
         inlineStyles: true,
     }