+++ packages/plugin-packages/mashroom-portal/test/backend/controllers/PortaHttpProxyController.test.ts
@@ -16,8 +16,9 @@ setPortalPluginConfig({
         restrictToRoles: [],
     },
     ssrConfig: {
-        ssrEnabled: false,
+        ssrEnable: false,
         renderTimoutMs: 2000,
+        cacheEnable: false,
         cacheTTLSec: 300,
         inlineStyles: true,
     }