+++ packages/plugin-packages/mashroom-portal/test/backend/utils/ssr_utils.test.ts
@@ -13,8 +13,9 @@ setPortalPluginConfig({
     autoExtendAuthentication: false,
     defaultProxyConfig: {},
     ssrConfig: {
-        ssrEnabled: true,
+        ssrEnable: true,
         renderTimoutMs: 2000,
+        cacheEnable: true,
         cacheTTLSec: 300,
         inlineStyles: true,
     }