+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 
 ## [unreleased]
 
+ * Portal: Fixed SSR cache key to avoid collisions
+ * Portal: Added possibility to turn of the SSR cache (even if @mashroom/mashroom-memory-cache is present)
  * OpenID Connect Security Provider: Removed the options *httpRequestRejectUnauthorized* and *httpRequestRetry* because they are no
    longer supported by [openid-client](https://github.com/panva/node-openid-client)
  * Vue Demo App: Upgrade to Vue3 and server-side rendering added